Fix matrix case05
